refactor: move routing types to Route.ts
