+++ app/frontend/src/types/Route.ts
@@ -1,0 +1,18 @@
+import AbstractView from "../views/AbstractView";
+
+type RouteGuardResult = true | false | string;
+export type RouteGuard = () => RouteGuardResult;
+
+export type RouteConfig = {
+  view: new () => AbstractView;
+  guard?: RouteGuard;
+};
+
+export type routeEvent = "nav" | "view";
+export type RouteChangeInfo = {
+  event: routeEvent;
+  from: string;
+  to: string;
+  view: AbstractView | null;
+};
+export type RouteChangeListener = (info: RouteChangeInfo) => void;
